Rename createDownloadTasks to declareSkiaTasks and move it to the SkikoProjectContext
