+++ skiko/buildSrc/src/main/kotlin/SkikoProjectContext.kt
@@ -1,3 +1,4 @@
+import de.undercouch.gradle.tasks.download.Download
 import org.gradle.api.DefaultTask
 import org.gradle.api.GradleException
 import org.gradle.api.Project
@@ -5,6 +6,7 @@ import org.gradle.api.provider.Provider
 import org.gradle.api.tasks.Copy
 import org.gradle.api.tasks.TaskProvider
 import org.gradle.api.tasks.bundling.Jar
+import org.gradle.kotlin.dsl.register
 import org.gradle.kotlin.dsl.withType
 import org.jetbrains.kotlin.gradle.dsl.KotlinMultiplatformExtension
 import java.io.File
@@ -25,6 +27,48 @@ class SkikoProjectContext(
 
     val allJvmRuntimeJars = mutableMapOf<Pair<OS, Arch>, TaskProvider<Jar>>()
 }
+
+fun SkikoProjectContext.declareSkiaTasks() {
+    val basicConfigs = listOf("android", "ios", "iosSim", "linux", "macos", "tvos", "tvosSim", "wasm", "windows")
+
+    basicConfigs.forEach { config ->
+        (if (config == "wasm") listOf("wasm") else listOf("arm64", "x64")).forEach { arch ->
+            val taskNameSuffix = joinToTitleCamelCase(config, arch)
+            val target = "$config-$arch"
+            val skiaReleaseTag = project.property("dependencies.skia.$target") as String
+
+            val skiaBaseUrl = "https://github.com/JetBrains/skia-pack/releases/download/$skiaReleaseTag"
+
+            val artifactId = "Skia-${skiaReleaseTag}-${config}-$buildType-${arch}"
+
+            val downloadSkiaTask = project.tasks.register<Download>("downloadSkia$buildType$taskNameSuffix") {
+                group = "Skia Binaries"
+
+                val skiaUrl = "$skiaBaseUrl/$artifactId.zip"
+                description = "downloads $skiaUrl"
+
+                onlyIfModified(true)
+                src(skiaUrl)
+                dest(skiko.dependenciesDir.resolve(
+                    "skia/$skiaReleaseTag/Skia-$skiaReleaseTag-$config-Release-${arch}.zip")
+                )
+            }
+
+            project.tasks.register<Copy>("unzipSkia$buildType$taskNameSuffix") {
+                group = "Skia Binaries"
+
+                val outputDir = skiko.dependenciesDir.resolve("skia/$skiaReleaseTag/$artifactId")
+                description = "unzips to $outputDir"
+
+                dependsOn(downloadSkiaTask)
+                from(project.zipTree(downloadSkiaTask.get().dest))
+
+                into(outputDir)
+            }
+        }
+    }
+}
+
 
 /**
  * Do not call inside tasks.register or tasks.call callback
